+++ CHANGES.rst
@@ -33,6 +33,7 @@ Bugfixes
   thanks :user:`SegiNyn`)
 - Fix validation error when clearing a date field in the registration form (:pr:`6470`)
 - Fix access error when a manager registers a user in a private registration form (:pr:`6486`)
+- Improve color handling in badge designer (auto-add ``#`` for hex colors) (:pr:`6492`)
 
 Accessibility
 ^^^^^^^^^^^^^